Revert "Use base::ByteCount in base::SysInfo."
This reverts commit 1d9a1d3425f0ab9b5350a70203c03a05ea83eabc.

Reason for revert: https://crbug.com/438649892
    Test failure on `android-arm-tests`

Bug: 429140103
Original change's description:
> Use base::ByteCount in base::SysInfo.
>
> Use base::ByteCount as the return value of
> SysInfo::AmountOfPhysicalMemory() and
> SysInfo::AmountOfAvailablePhysicalMemory() and adjust call sites across
> the codebase. Note that other methods in base::SysInfo could also return
> base::Bytes, but they're not upgraded in this CL to avoid making it
> bigger.
>
> base::ByteCount is a class to make bytes manipulations type safe, easier
> to read and less error-prone.
>
> NO_IFTTT=Code style change in base/system/sys_info.cc, does not affect code behavior.
>
> AX-Relnotes: n/a
> Bug: 429140103
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6698912
> Owners-Override: Gabriel Charette <gab@chromium.org>
> Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1501122}

Bug: 429140103
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6849446
Auto-Submit: Peter Pakkenberg <pbirk@chromium.org>
Owners-Override: Peter Pakkenberg <pbirk@chromium.org>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1501264}

+++ base/features.cc
@@ -78,7 +78,7 @@ BASE_FEATURE(kFastFilePathIsParent,
 BASE_FEATURE(kLowEndMemoryExperiment,
              "LowEndMemoryExperiment",
              FEATURE_DISABLED_BY_DEFAULT);
-BASE_FEATURE_PARAM(int,
+BASE_FEATURE_PARAM(size_t,
                    kLowMemoryDeviceThresholdMB,
                    &kLowEndMemoryExperiment,
                    "LowMemoryDeviceThresholdMB",
